pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.116.2",
    "groq>=0.31.1",
    "jinja2>=3.1.6",
    "numpy>=2.3.3",
    "pdfminer-six>=20250506",
    "pydantic>=2.11.9",
    "python-multipart>=0.0.20",
    "scikit-learn>=1.7.2",
    "scipy>=1.16.2",
    "uvicorn[standard]>=0.36.0",
]
